workshop01/public_private_channels/pyproject.toml
[tool.poetry]
name = "public-private-channels"
version = "0.1.0"
description = "Practice problem on websocket"
authors = ["Md. Rifaet - Ullah Rifat <md.rifaetullahrifat@gmail.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
django = "^5.1.2"
channels = {extras = ["daphne"], version = "^4.1.0"}
environs = "^11.0.0"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
